requirements/dev.txt
-r base.txt

fabric3
tox
coverage
isort
pycodestyle
pylint
ipython
ipdb
pudb
